[REXA-22] MR : update after comments on MR
