+++ frontend/src/app/Header.js
@@ -9,7 +9,7 @@ import { performLogout } from "../auth/redux/authDuck"
 import { ACCESS_TOKEN } from "../constants"
 import { toast } from "react-toastify"
 
-import style from "../common/common.module.scss";
+import style from "../common/common.module.scss"
 
 const Header = (props) => {
   const [expanded, setExpanded] = useState(false)
@@ -75,32 +75,35 @@ const Header = (props) => {
                 </NavText>
               </NavItem>
 
-                            <NavItem eventKey="rexa/settings" active={props.location.pathname.startsWith("/rexa/settings")}>
-                                <NavIcon>
-                                    <i
-                                        className="fa fa-cogs"
-                                        aria-hidden="true"
-                                        style={{fontSize: "1.75em"}}
-                                    />
-                                </NavIcon>
-                                <NavText>
-                                    <Link to="rexa/project">Project View</Link>
-                                </NavText>
-                            </NavItem>
+              <NavItem
+                eventKey="rexa/settings"
+                active={props.location.pathname.startsWith("/rexa/settings")}
+              >
+                <NavIcon>
+                  <i
+                    className="fa fa-cogs"
+                    aria-hidden="true"
+                    style={{ fontSize: "1.75em" }}
+                  />
+                </NavIcon>
+                <NavText>
+                  <Link to="rexa/project">User settings</Link>
+                </NavText>
+              </NavItem>
 
-                            <NavItem onClick={handleLogOut} eventKey="logout">
-                                <NavIcon>
-                                    <i className="fa fa-sign-out" style={{fontSize: "1.75em"}}/>
-                                </NavIcon>
-                                <NavText>Logout</NavText>
-                            </NavItem>
-                        </SideNav.Nav>
-                    </SideNav>
-                </ClickOutside>
-            )}
-        </>
-    )
-};
+              <NavItem onClick={handleLogOut} eventKey="logout">
+                <NavIcon>
+                  <i className="fa fa-sign-out" style={{ fontSize: "1.75em" }} />
+                </NavIcon>
+                <NavText>Logout</NavText>
+              </NavItem>
+            </SideNav.Nav>
+          </SideNav>
+        </ClickOutside>
+      )}
+    </>
+  )
+}
 
 Header.propTypes = {
   authenticated: PropTypes.bool.isRequired,
