+++ src/main/resources/db/migration/V2_2__data_user_settings_.sql
@@ -1,2 +1,3 @@
 -- V2_2__data_user_settings_.sql
+-- mdp : admin
 INSERT INTO rexa.user_settings VALUES ('admin', '$2y$10$g8EqWNrFpzPZa2R7NAl2le8yaXFzeBX2piZhM5MFF6pJlmVQn7s2C', 'http://localhost','910669d4-18e6-447e-8baa-269c104a6925');
